feat: use miette for diagnostics
