+++ crates/package_json/Cargo.toml
@@ -12,6 +12,7 @@ repository.workspace = true
 
 [dependencies]
 thiserror = { workspace = true }
+miette = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
