+++ crates/package_manager/Cargo.toml
@@ -19,6 +19,7 @@ pacquet_tarball = { workspace = true }
 async-recursion = { workspace = true }
 elsa = { workspace = true }
 futures-util = { workspace = true }
+miette = { workspace = true }
 reqwest = { workspace = true }
 reqwest-middleware = { workspace = true }
 reqwest-retry = { workspace = true }
