+++ crates/registry/Cargo.toml
@@ -19,6 +19,7 @@ futures-util = { workspace = true }
 reqwest = { workspace = true }
 reqwest-middleware = { workspace = true }
 reqwest-retry = { workspace = true }
+miette = { workspace = true }
 node-semver = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
