+++ crates/tarball/Cargo.toml
@@ -15,6 +15,7 @@ pacquet_cafs = { workspace = true }
 
 bytes = { workspace = true }
 futures-util = { workspace = true }
+miette = { workspace = true }
 reqwest = { workspace = true }
 ssri = { workspace = true }
 tar = { workspace = true }
